F1-1: Move base models to own folder
